+++ skysql_chain.pem
@@ -1,0 +1,48 @@
+-----BEGIN CERTIFICATE-----
+MIIELzCCAxegAwIBAgIUNz1kWFjbLVDWVCBnEAsf95UPMS8wDQYJKoZIhvcNAQEL
+BQAwQTEQMA4GA1UEChMHTWFyaWFEQjEPMA0GA1UECxMGU2t5U1FMMRwwGgYDVQQD
+ExNyb290LXBraS5za3lzcWwubmV0MB4XDTE5MDkwNTE4MjMxMVoXDTI5MDkwMjE4
+MjM0MFowQTEQMA4GA1UEChMHTWFyaWFEQjEPMA0GA1UECxMGU2t5U1FMMRwwGgYD
+VQQDExNyb290LXBraS5za3lzcWwubmV0MIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8A
+MIIBCgKCAQEA0orsl93gB0MZuiby/4QmcmK2OrS/hEcDI/AN+Dpn9c3JKeEVp2OD
+sft47OHGwdgyagBtNV6zZgOc6IOnwt+rGDmrmiuxHkf/XWV+y66skWAtMyM7ycCL
+J3z5dO6xaZvYKJyhPcnx2NROEAJrkdVfoyJCtCElDDdRrknXWLPfZrph8E7I2mDP
+SV8ZF4wdxbU7oHKM4CoTRgXQnCDq2Wv8OLZr4Mq224nSmEJK+cXRwKqbFUvuiSco
+bTBnJjyeKldqJ/lCRwu9fU6fBHFuBNUEvZBzavt0B8SYi/l22wYHxlpOslowTaG4
+Lh8Nj79PP7rsy44hHvOBGc/ZsKIGCDOIMwIDAQABo4IBHTCCARkwDgYDVR0PAQH/
+BAQDAgEGMA8GA1UdEwEB/wQFMAMBAf8wHQYDVR0OBBYEFHAIHx8QSWENuYb/xmuh
+17S2nnXWMB8GA1UdIwQYMBaAFHAIHx8QSWENuYb/xmuh17S2nnXWMEAGCCsGAQUF
+BwEBBDQwMjAwBggrBgEFBQcwAoYkaHR0cDovLzEyNy4wLjAuMTo4MjAwL3YxL3Br
+aV9yb290L2NhMB4GA1UdEQQXMBWCE3Jvb3QtcGtpLnNreXNxbC5uZXQwHAYDVR0e
+AQH/BBIwEKAOMAyCCnNreXNxbC5uZXQwNgYDVR0fBC8wLTAroCmgJ4YlaHR0cDov
+LzEyNy4wLjAuMTo4MjAwL3YxL3BraV9yb290L2NybDANBgkqhkiG9w0BAQsFAAOC
+AQEAfwo8ZW666UjHJ4DY+M9tDgRwFwFd7v3EBhLrGvkD+CWaiJIS9RnWwE0gn9SU
+syBvRn3PrjsveCR3cIjqAzUplOyMMvvJ77E8rzfQEwOhHbATyKNQG32KaitCdEBP
+v0XDb7SBw2eKQxdahMcT5yxh9DkCizTXE8usZIiW+V9FVcEPPNia4d9ZMlmLWMcP
+pZlxE4W5ngU6iCN7PJ3aeKrk4Y1PM36XJ11f5pouMULUvqbjepa/R1KJt27OSbrJ
+RjHDa+s0AljgPZDl7KqQOOA5hrNT1Om+5IVs+uAbY7mWQC2GwYlFsg5laqWf7SC0
+hPvVLDb8GaRK6LA4PCROZwiM9g==
+-----END CERTIFICATE-----
+-----BEGIN CERTIFICATE-----
+MIID4jCCAsqgAwIBAgIUB9abD+hn9+dtYuqwkA3Rj5ZrPgUwDQYJKoZIhvcNAQEL
+BQAwQTEQMA4GA1UEChMHTWFyaWFEQjEPMA0GA1UECxMGU2t5U1FMMRwwGgYDVQQD
+ExNyb290LXBraS5za3lzcWwubmV0MB4XDTE5MDkwNTE4MjMxOFoXDTI5MDkwMjE4
+MjM0OFowGTEXMBUGA1UEAxMOcGtpLnNreXNxbC5uZXQwggEiMA0GCSqGSIb3DQEB
+AQUAA4IBDwAwggEKAoIBAQDYov+F2ijXdIiZ0AuX4fAJ6KQ16zb4mQ2qgsrO02yW
+kF3EJV6/XQO0WqGok4SjcvLBLuSsQBFahtgB70d/YZ+PBUrwzzmgWa3Ga+GuzKl6
+O2QI8vu7l8D0esJe7mY4KsAwNIvMUAdqUUCgB01KmCIwWoVqN1h65dX1qOf1N6qk
+f+rXeFKBGoDq/DM6zR90irpYBt2guE5iZd5r63JMXANlmh44IWxEswBqAa4B+GlY
+7m7Psk9E+i4rexN45+815SLnHr86y3PNUlFfzgfQwXCLVPqSPGfNzyEz8MZUVYQv
+zg+hZmrTe9nCekMi3hk97Dh1x40Y1rSTtbQjUu2SvJepAgMBAAGjgfkwgfYwDgYD
+VR0PAQH/BAQDAgEGMA8GA1UdEwEB/wQFMAMBAf8wHQYDVR0OBBYEFLwrAxEtsiHF
+Oz/dRW65RxwIHHVmMB8GA1UdIwQYMBaAFHAIHx8QSWENuYb/xmuh17S2nnXWMEAG
+CCsGAQUFBwEBBDQwMjAwBggrBgEFBQcwAoYkaHR0cDovLzEyNy4wLjAuMTo4MjAw
+L3YxL3BraV9yb290L2NhMBkGA1UdEQQSMBCCDnBraS5za3lzcWwubmV0MDYGA1Ud
+HwQvMC0wK6ApoCeGJWh0dHA6Ly8xMjcuMC4wLjE6ODIwMC92MS9wa2lfcm9vdC9j
+cmwwDQYJKoZIhvcNAQELBQADggEBAAZZKyFT+mVwuafkBOBYqXb/dCPdqbUnGBic
+E2dRK0sxKYbeq7I3lo95UXrtfNBEMY740ZJzUwi6whDUMGNMoV0yFRPHPYvmopC5
+wCUA62pPuvHEqwo7HSuO3TBmt5x0b2e9R0gJ535GZSTQI+ArseUwn5IJ2v/BUIRJ
+xjAMwRmM9TOWcK6VLEBZoHXEzENrBLHr0fKhVyJhhuQV+xeEVY28odwzwH85AyUk
+1lrAIzuz3YCDHtjL449U+hdz/2tytI1KXJscm/mrAhtgUjQnKCY0fFkJESL+TDwX
+Sdb13rs8ZQvwanpcOt+Kg86O/vz2P5JLC8fK1L4aUilt0X5b8gc=
+-----END CERTIFICATE-----
